TASK-6 refactoring
