+++ backend/src/main/java/ru/hh/school/adaptation/dto/WorkflowStepDto.java
