+++ backend/src/main/java/ru/hh/school/adaptation/entities/WorkflowStep.java
@@ -1,0 +1,42 @@
+package ru.hh.school.adaptation.entities;
+
+import javax.persistence.*;
+
+@Entity
+@Table(name = "workflow_leonid")
+public class Workflow {
+
+  @Id
+  @Column(name = "id")
+  private Integer id;
+
+  @Column(name = "name", nullable = false)
+  private String name;
+
+  @Column(name = "description")
+  private String description;
+
+  public Integer getId() {
+    return id;
+  }
+
+  public void setId(Integer id) {
+    this.id = id;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public void setName(String name) {
+    this.name = name;
+  }
+
+  public String getDescription() {
+    return description;
+  }
+
+  public void setDescription(String description) {
+    this.description = description;
+  }
+}
